+++ Str_Generation/LammpsStrBCC.m
@@ -1,13 +1,32 @@
-function [varargout] = LammpsStrBCC(num_cell,lattice_const,atom_type,atom_charge)
-
-% function [coord_str,size_box] = LammpsStrBCC(num_cell,lattice_const)
+function [varargout] = LammpsStrBCC(varargin)
+%% Description
+% function [varargout] = LammpsStrBCC(num_cell,lattice_const,atom_type,atom_charge)
 %
 % Input:
+%
 % num_cell      :   3*1 vector. Each elements represents # of cells in this direction
+%
 % lattice_const :   lattice constant of struture.
+%
+% atom_type     :   *Case 1* : atom_type = 1, all atom will be set to type 1;
+%                   *Case 2* : atom_type = 2, each atom in cell will get an
+%                   unique type
+%                   *Case 3* : atom_type is matric. each of atom will be set
+%                   to type correspond number in matric. Matric lenghth
+%                   should be equal to # of atoms in cell strickly.
+%
+% atom_charge   :   *Case 1* : atom_charge = 0; All atom charge will be set
+%                   to 0.
+%                   *Case 2* : atom_charge single number. All atom will be set 
+%                   to the same charge.
+%                   *Case 3* : atom_charge is a matric. Each atom will be set
+%                   to a charge correspond number in matric. Same as Case 3
+%                   of atom_type, length of matric should be equal to
+%                   number of atoms in cell
 
+%% Reading Input
 
-num_cell_tot        =   num_cell(1) * num_cell(2) * num_cell(3);
+num_cell_tot        =   varargin{1}(1) * varargin{1}(2) * varargin{1}(3);
 
 str_mtr             =   [
                         0 0 0
@@ -16,40 +35,42 @@ str_mtr             =   [
 
 [num_cell_atom,num_dim]       =   size(str_mtr);
 
-if atom_type == 0
-    atom_type = ones(num_cell_atom,1);
+if varargin{3} == 1
+    varargin{3} = ones(num_cell_atom,1);
     num_atom_type = 1;
-elseif atom_type == 2;
-    atom_type = 1:num_cell_atom;
+elseif varargin{3} == 2;
+    varargin{3} = 1:num_cell_atom;
     num_atom_type = num_cell_atom;
 else
-    num_atom_type = length(atom_type);
+    num_atom_type = length(varargin{3});
 end
 
-if atom_charge == 0
-    atom_charge = zeros(num_cell_atom,1);
-elseif length(atom_charge) == 1
-    atom_charge = ones(num_cell_atom,1) * atom_charge;
+if varargin{4} == 0
+    varargin{4} = zeros(num_cell_atom,1);
+elseif length(varargin{4}) == 1
+    varargin{4} = ones(num_cell_atom,1) * varargin{4};
 end
 
 atom_style      =   'full';
 
 fprintf("# of atoms: %d\n",num_cell_tot*num_cell_atom)
 
+%% Writing Data File
+
 for dim = 1 : num_dim
-    size_box(dim,:) = [0 , lattice_const * num_cell(dim)];
+    size_box(dim,:) = [0 , varargin{2} * varargin{1}(dim)];
 end
 
-for x = 1 : num_cell(1)
-    for y = 1 : num_cell(2)
-        for z = 1 : num_cell(3)
-            cell_now = z + num_cell(3) * (y-1) + num_cell(3) * num_cell(2) * (x-1);
+for x = 1 : varargin{1}(1)
+    for y = 1 : varargin{1}(2)
+        for z = 1 : varargin{1}(3)
+            cell_now = z + varargin{1}(3) * (y-1) + varargin{1}(3) * varargin{1}(2) * (x-1);
             for atom = 1 : num_cell_atom
                 data_str(cell_now,atom,1) = (cell_now - 1) * num_cell_atom + atom;
                 data_str(cell_now,atom,2) = cell_now;
-                data_str(cell_now,atom,3) = atom_type(atom);
-                data_str(cell_now,atom,4) = atom_charge(atom);
-                data_str(cell_now,atom,5:7) = [x-1, y-1, z-1] * lattice_const + 0.5 * lattice_const * str_mtr(atom,:);
+                data_str(cell_now,atom,3) = varargin{3}(atom);
+                data_str(cell_now,atom,4) = varargin{4}(atom);
+                data_str(cell_now,atom,5:7) = [x-1, y-1, z-1] * varargin{2} + 0.5 * varargin{2} * str_mtr(atom,:);
             end
         end
     end